Extract ConfirmationsEdit component
